requirements.txt
neo4j
openai
pinecone-client
pyvis
networkx
tqdm
python-dotenv